requirements.txt
flask>=3.0.3
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.7
flask-limiter>=3.8.0
flask-wtf>=1.2.2
flask-talisman>=1.1.0   # opcional (CSP/headers extras)

google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.149.0

openai>=1.44.0
requests>=2.32.3
python-dotenv>=1.0.1

gunicorn>=23.0.0
psycopg2-binary>=2.9.9

selenium>=4.25.0
chromedriver-autoinstaller>=0.6.4

APScheduler>=3.11.0
redis>=5.0.8
sentry-sdk[flask]>=2.13.0

cryptography>=43.0.1
pillow>=10.4.0
pandas>=2.2.2
matplotlib>=3.9.2
wordcloud>=1.9.3
fpdf2>=2.8.2
pytz>=2024.1
seaborn
google-generativeai
